refactor(gigs): rename queries file
